Extract growl_fallback.py
