+++ .gitignore
@@ -1,1 +1,2 @@
 *.pyc
+/pytest_growl.egg-info
